Refactor, move OD models to separate module, fix YOLO prediction generation
